+++ examples/partial/main.tf
@@ -13,12 +13,11 @@ module "datarobot_infra" {
   name        = local.name
   domain_name = "${local.name}.yourdomain.com"
 
-  existing_vpc_id                      = "vpc-1234556abcdef"
-  existing_kubernetes_nodes_subnet_ids = ["subnet-abcde012", "subnet-bcde012a", "subnet-fghi345a"]
-  existing_private_route53_zone_id     = "Z06110132R7HO9BLI64XY"
-  existing_acm_certificate_arn         = "arn:aws:acm:us-west-2:000000000000:certificate/00000000-0000-0000-0000-000000000000"
-  existing_kms_key_arn                 = "arn:aws:kms:us-west-2:500395161552:key/00000000-0000-0000-0000-000000000000"
-  existing_s3_bucket_id                = "datarobot-file-storage-12345"
+  existing_vpc_id                  = "vpc-1234556abcdef"
+  existing_kubernetes_node_subnets = ["subnet-abcde012", "subnet-bcde012a", "subnet-fghi345a"]
+  existing_private_route53_zone_id = "Z06110132R7HO9BLI64XY"
+  existing_acm_certificate_arn     = "arn:aws:acm:us-west-2:000000000000:certificate/00000000-0000-0000-0000-000000000000"
+  existing_s3_bucket_id            = "datarobot-file-storage-12345"
 
   # disable public internet access to the Kubernetes API endpoint
   kubernetes_cluster_endpoint_public_access = false