+++ modules/aws-ebs-csi-driver/variables.tf
@@ -8,11 +8,6 @@ variable "kubernetes_cluster_name" {
   type        = string
 }
 
-variable "aws_ebs_csi_kms_arn" {
-  description = "ARN of the KMS key used to encrypt EBS volumes"
-  type        = string
-}
-
 variable "custom_values_templatefile" {
   description = "Custom values templatefile to pass to the helm chart"
   type        = string