+++ modules/external-dns/main.tf
@@ -1,6 +1,6 @@
 module "external_dns_pod_identity" {
   source  = "terraform-aws-modules/eks-pod-identity/aws"
-  version = "~> 1.0"
+  version = "~> 2.0"
 
   name = "external-dns"
 
@@ -21,9 +21,9 @@ module "external_dns_pod_identity" {
 resource "helm_release" "external_dns" {
   name       = "external-dns"
   namespace  = "external-dns"
-  repository = "https://charts.bitnami.com/bitnami"
+  repository = "https://kubernetes-sigs.github.io/external-dns"
   chart      = "external-dns"
-  version    = "8.5.1"
+  version    = "1.19.0"
 
   create_namespace = true