+++ modules/ingress-nginx/main.tf
@@ -3,7 +3,7 @@ resource "helm_release" "ingress_nginx" {
   namespace  = "ingress-nginx"
   repository = "https://kubernetes.github.io/ingress-nginx"
   chart      = "ingress-nginx"
-  version    = "4.11.5"
+  version    = "4.13.2"
 
   create_namespace = true